+++ flywheel/__init__.py
@@ -1,7 +1,7 @@
 """ flywheel """
 from dynamo3 import (CheckFailed, ConditionalCheckFailedException, Binary,
                      STRING, NUMBER, BINARY, STRING_SET, NUMBER_SET,
-                     BINARY_SET)
+                     BINARY_SET, Limit)
 
 from .fields import Field, Composite, GlobalIndex
 from .fields.types import set_, TypeDefinition
@@ -9,4 +9,4 @@ from .models import Model
 from .engine import Engine
 from .query import DuplicateEntityException, EntityNotFoundException
 
-__version__ = '0.4.3'
+__version__ = '0.4.10'