+++ CHANGES.rst
@@ -7,6 +7,35 @@ Changelog
 * Flywheel no longer forces raise_on_conflict to be True when you sync changes to fields that are part of a composite field. It is now up to the user to avoid putting their composite fields into an inconsistent state.
 * Feature: sync() has a new argument, ``no_read``, which changes the behavior for syncing models with no changes. Instead of performing a GET, it will leave them as-is. This should make it easer to perform batch syncs without worrying as much about wasted bandwidth on GETs. 
 
+0.4.10
+------
+* Feature: Add ``exists()`` method to Engine (:issue:`45`)
+
+0.4.9
+-----
+* Feature: Add ``save()`` method to Models (:issue:`40`)
+* Feature: Add ``update_field()`` method to Engine (:issue:`43`)
+
+0.4.8
+-----
+* Bug fix: Bad function call in ``index_pk_dict_``
+
+0.4.7
+-----
+* New ``index_pk_dict_`` method for constructing `exclusive_start_key` for index queries (:issue:`34`)
+
+0.4.6
+-----
+* Pass exclusive_start_key through to dynamo3 (:issue:`34`)
+
+0.4.5
+-----
+* Bug fix: Calling refresh() could sometimes crash from unordered results.
+
+0.4.4
+-----
+* Bug fix: Mutable field defaults are no longer shared among model instances
+
 0.4.3
 -----
 * Bug fix: Incorrect ``ConditionalCheckFailedException`` when syncing changes to a Composite field.
@@ -33,7 +62,7 @@ Changelog
 * Feature: Fields support multiple validation checks
 * Feature: Fields have an easy way to enforce non-null values (``nullable=False``)
 
-Data type changes are due to an `update in the DynamoDB API
+Data type changes are due to `an update in the DynamoDB API
 <https://aws.amazon.com/blogs/aws/dynamodb-update-json-and-more/>`_
 
 0.3.0